Ressource dynamischer gestaltet

+++ 01. Eingerichtet/Ressource.h
@@ -12,17 +12,17 @@ private:
 	static int typc;				//Attribute der Klasse
 	int typ;
 	string name;
-	float menge;
+	int menge;
 
 public:
 
 	~Ressource();					//Destruktor und Konstruktor
-	Ressource(string, float);
-	float getmenge();				//Get und Set Methoden
+	Ressource(string, int);
+	int getmenge();				//Get und Set Methoden
 	string getname();
 	void setname(string);
-	void setmenge(float);
-	void addmenge(float);			//Zum Ändern der Menge
+	void setmenge(int);
+	void addmenge(int);			//Zum Ändern der Menge
 	void darstellen(RenderWindow*); // Ausgabe der Werte in der oberen linken Ecke
 
 };
